refactor: consolidate mixin modules into _mixins/ package
Move 7 ToolRegistry mixin modules from package root (_admin.py,
_callbacks.py, etc.) into a dedicated _mixins/ subdirectory to
reduce top-level clutter. All relative imports updated to use
parent-package references.
