+++ src/toolregistry/tool_registry.py
@@ -26,13 +26,15 @@ from .types import (
 from .events import ChangeCallback, ChangeEvent, ChangeEventType
 from .tool_discovery import TOOL_DISCOVERY_NAME, ToolDiscoveryTool
 
-from ._admin import AdminMixin
-from ._callbacks import ChangeCallbackMixin
-from ._enable_disable import EnableDisableMixin
-from ._logging import ExecutionLoggingMixin
-from ._namespace import NamespaceMixin
-from ._permissions import PermissionsMixin
-from ._registration import RegistrationMixin
+from ._mixins import (
+    AdminMixin,
+    ChangeCallbackMixin,
+    EnableDisableMixin,
+    ExecutionLoggingMixin,
+    NamespaceMixin,
+    PermissionsMixin,
+    RegistrationMixin,
+)
 
 logger = logging.getLogger(__name__)
